Fix safe deletion of unsaved environments in web_handlers

+++ py/visdom/server/handlers/web_handlers.py
@@ -415,15 +415,22 @@ class DeleteEnvHandler(BaseHandler):
         self.env_path = app.env_path
         self.login_enabled = app.login_enabled
 
-    @staticmethod
-    def wrap_func(handler, args):
-        eid = extract_eid(args)
-        if eid is not None:
+@staticmethod
+def wrap_func(handler, args):
+    eid = extract_eid(args)
+    if eid is not None:
+
+        # Safe delete from state
+        if eid in handler.state:
             del handler.state[eid]
-            if handler.env_path is not None:
-                p = os.path.join(handler.env_path, "{0}.json".format(eid))
+
+        # Safe file deletion
+        if handler.env_path is not None:
+            p = os.path.join(handler.env_path, "{0}.json".format(eid))
+            if os.path.exists(p):
                 os.remove(p)
-            broadcast_envs(handler)
+
+        broadcast_envs(handler)
 
     @check_auth
     def post(self):
